Create bash script for CircleCI to run.
The reason behind creating this bash script is that not all languages have a CIMG that the CI can pull.
I'm also not at a skill level where I know how to use variables in a CircleCI config thus, was unable to have the flexibility of a for loop to loop over algorithm names.
This way, I'm ensuring that the CI will be able to format, test, and run all algorithm implementations in all languages.